+++ core/Http/Server/Server.php
@@ -14,14 +14,18 @@ use core\Server\Server as BaseServer;
 class Server
 {
     private static mixed $server; // 服务套接字
-    private static array $clients = array(); // 客户端套接字
-    private static array $handlers; // 消费者列表
-    private static array $tasks;
+    private static mixed $eventServer; // Event套接字
+    private static array $tasks = array();    // 代办事项
+    private static array $transfer = array(); // 客户端套接字
+    private static array $handlerScokets = array(); // 消费者套接字
+    private static array $handlerFifos = array();
 
     /**
      * Summary of launch
      *
      * @return void
+     * @throws \Exception
+     * @throws \Exception
      */
     public static function launch(): void
     {
@@ -34,47 +38,51 @@ class Server
             try {
                 // 创建连接
                 self::$server = socket_create(AF_INET, SOCK_STREAM, SOL_TCP);
+                self::$eventServer = socket_create(AF_INET, SOCK_STREAM, SOL_TCP);
+
+                // 端口非互斥模式
                 socket_set_option(self::$server, SOL_SOCKET, SO_REUSEADDR, 1);
                 socket_set_option(self::$server, SOL_SOCKET, SO_REUSEPORT, 1);
+                socket_set_option(self::$eventServer, SOL_SOCKET, SO_REUSEPORT, 1);
+                socket_set_option(self::$eventServer, SOL_SOCKET, SO_REUSEPORT, 1);
 
                 // 监听指定HTTP端口
                 socket_bind(self::$server, '127.0.0.1', Config::get('http.server_port'));
+                socket_bind(self::$eventServer, '127.0.0.1', 2787);
+
                 // 监听连接
-                if (socket_listen(self::$server) === false) {
+                if (!socket_listen(self::$server) || !socket_listen(self::$eventServer)) {
                     throw new \Exception(socket_strerror(socket_last_error()));
                 }
             } catch (\Exception $e) {
+                socket_close(self::$server);
+                socket_close(self::$eventServer);
                 echo '发生错误: ' . $e->getMessage() . PHP_EOL;
                 return;
             }
-
-            // 初始化进程管理器
 
             // 启动消费者
             $handlerProcessIds = array();
             for ($i = 0; $i < Config::get('http.server_handle_count'); $i++) {
                 $event = EventHandler::create();
-                self::$handlers[] = $event;
+                self::$handlerFifos[] = $event;
                 $handlerProcessIds[] = $event->pid;
             }
             Process::guard();
-            $server->info(['handlerProcessIds' => $handlerProcessIds, 'serverProcessId' => posix_getpid(),]);
+            $server->info(['handlerProcessIds' => $handlerProcessIds, 'serverProcessId' => posix_getpid()]);
 
             // 开始循环监听
             while (true) {
                 // 监听服务端和客户端写入
-                $readList = array_merge([self::$server], array_column(self::$clients, 'socket'));
+                $readList = array_merge([self::$server], [self::$eventServer], self::$handlerScokets, array_column(self::$transfer, 'socket'),);
 
                 // 只监听可以写入缓冲区的客户端
                 $writeList = [];
 
                 // 监听所有客户端的异常信息
-                $exceptList = array_column(self::$clients, 'socket');
-
-                echo '开始监听' . PHP_EOL;
+                $exceptList = array_column(self::$transfer, 'socket');
 
                 if (socket_select($readList, $writeList, $exceptList, null) !== false) {
-                    echo '收到数据' . PHP_EOL;
                     // 处理异常连接
                     foreach ($exceptList as $socket) {
                         if (empty(socket_read($socket, 1024))) {
@@ -85,7 +93,6 @@ class Server
                                 unset($readList[$index]);
                             }
                         }
-                        echo '有客户端异常' . PHP_EOL;
                     }
 
                     // 处理可读消息
@@ -97,10 +104,19 @@ class Server
                             self::addClient($client);
 
                             // 来自客户端的消息
+                        } elseif (self::$eventServer === $socket) {
+                            $handler = socket_accept($socket);
+                            self::$handlerScokets[] = $handler;
                         } else {
                             echo '客户端发来数据' . PHP_EOL;
+                            $socketName = spl_object_hash($socket);
+
+                            // 取头消息
                             $context = socket_read($socket, 1);
-                            // 来自消费者的消息
+                            if ($context === '') {
+                                self::removeClient($socket);
+                            }
+
                             if ($context === '@') {
                                 echo '是消费者消息' . PHP_EOL;
                                 $content = '';
@@ -110,14 +126,10 @@ class Server
                                     if (!isset($_[1])) {
                                         $symbol = socket_read($socket, 1);
                                         if ($symbol === '#') {
-                                            if ($content === 'stop') {
-                                                foreach (array_column(self::$clients, 'socket') as $item) {
+                                            if ($content === 'stop')
+                                                self::release();
 
-                                                    self::removeClient($item);
-                                                }
-                                                socket_close(self::$server);
-                                                exit;
-                                            }
+                                            // 存如信息并清空暂存区
                                             $_[] = $content;
                                             $content = '';
                                             continue;
@@ -126,12 +138,14 @@ class Server
                                         $content .= $symbol;
                                     } else {
                                         $_[] = socket_read($socket, intval($_[1]));
-                                        //handler
 
-                                        if (isset(self::$tasks[$_[0]])) {
-                                            socket_write(self::$tasks[$_[0]]['socket'], $_[2]);
-                                            self::$clients[$_[0]] = array('socket' => self::$tasks[$_[0]]['socket'], 'context' => false, 'header_context' => '', 'body_context' => '', 'method' => 'undefined', 'path' => '', 'version' => '', 'data' => '', 'header' => array(), 'complete' => false, 'createTime' => time(), 'failedCount' => 0,);
-                                            unset(self::$tasks[$_[0]]);
+                                        $nominator = $_[0];
+                                        $context = $_[2];
+
+                                        if ($client = self::$tasks[$nominator] ?? null) {
+                                            socket_write($client['socket'], $context);
+                                            self::addClient($client['socket']);
+                                            unset(self::$tasks[$nominator]);
                                         }
                                         break;
                                     }
@@ -140,14 +154,11 @@ class Server
                             }
 
                             // 新的客户端进入
-                            $client = &self::$clients[spl_object_hash($socket)];
+                            $client = &self::$transfer[$socketName];
                             // 读取数据
                             $context .= socket_read($socket, 1024);
                             $client['context'] .= $context;
 
-                            if ($context === '') {
-                                self::removeClient($socket);
-                            }
                             if ($client['method'] === 'undefined') {
                                 if (str_contains($client['context'], "\r\n\r\n")) {
                                     $_ = explode("\r\n\r\n", $client['context']);
@@ -196,22 +207,55 @@ class Server
                             }
 
                             if ($client['complete'] === true) {
-                                echo '当前客户端总数' . count(self::$clients) . PHP_EOL;
+                                echo '当前客户端总数' . count(self::$transfer) . PHP_EOL;
                                 // 推送处理
-                                self::$tasks[spl_object_hash($socket)] = $client;
-                                self::$handlers[array_rand(self::$handlers)]->push(spl_object_hash($socket), $client);
-                                unset(self::$clients[spl_object_hash($socket)]);
+                                self::$tasks[$socketName] = $client;
+                                $handle = self::$handlerFifos[array_rand(self::$handlerFifos)];
+                                $handle->push($socketName, $client);
+                                unset(self::$transfer[$socketName]);
                             }
                         }
                     }
                 }
             }
-            // });
-            // sleep(1);
-            // Process::guard();
         }
     }
 
+    private static function removeClient($socket): void
+    {
+        echo '客户端' . spl_object_hash($socket) . '断开连接' . PHP_EOL;
+        // 移除客户端
+        if (self::$transfer[spl_object_hash($socket)]) {
+            socket_close($socket);
+            unset(self::$transfer[spl_object_hash($socket)]);
+        }
+    }
+
+    private static function addClient($socket): void
+    {
+        self::$transfer[spl_object_hash($socket)] = array('socket' => $socket, 'context' => false, 'header_context' => '', 'body_context' => '', 'method' => 'undefined', 'path' => '', 'version' => '', 'data' => '', 'header' => array(), 'complete' => false, 'createTime' => time(), 'failedCount' => 0,);
+    }
+
+    /**
+     * @return void
+     */
+    private static function release(): void
+    {
+        foreach (array_column(self::$transfer, 'socket') as $item) {
+            self::removeClient($item);
+        }
+        foreach (self::$eventServer as $item) {
+            $item->push('stop', []);
+            $item->close();
+        }
+        socket_close(self::$server);
+        socket_close(self::$eventServer);
+        exit;
+    }
+
+    /**
+     * @throws \Exception
+     */
     public static function stop(): void
     {
         if ($server = BaseServer::load('HTTP_SERVER')) {
@@ -227,37 +271,28 @@ class Server
                 $server->release();
                 exit;
             }
-            Process::initialization();
-            $info = $server->info();
-            foreach ($info['handlerProcessIds'] as $pid) {
-                Process::kill($pid);
+            if (Process::initialization()) {
+                // 创建客户端套接字
+                $socket = socket_create(AF_INET, SOCK_STREAM, SOL_TCP);
+
+                // 连接服务器
+                socket_connect($socket, '127.0.0.1', Config::get('http.server_port'));
+
+                socket_write($socket, '@stop#');
+                socket_close($socket);
+
+                $info = $server->info();
+                foreach ($info['handlerProcessIds'] as $pid) {
+                    Process::kill($pid);
+                }
+                Process::kill($info['serverProcessId']);
+
+                $server->release();
+                Console::pgreen('Http Server stop success!');
             }
-            Process::kill($info['serverProcessId']);
-            // 创建客户端套接字
-            $socket = socket_create(AF_INET, SOCK_STREAM, SOL_TCP);
 
-            // 连接服务器
-            socket_connect($socket, '127.0.0.1', Config::get('http.server_port'));
 
-            socket_write($socket, '@stop#');
-            socket_close($socket);
-
-            $server->release();
-            Console::pgreen('Http Server stop success!');
-            var_dump($info);
+            //            var_dump($info);
         }
-    }
-
-    private static function removeClient($socket): void
-    {
-        echo '客户端' . spl_object_hash($socket) . '断开连接' . PHP_EOL;
-        // 移除客户端
-        socket_close($socket);
-        unset(self::$clients[spl_object_hash($socket)]);
-    }
-
-    private static function addClient($socket): void
-    {
-        self::$clients[spl_object_hash($socket)] = array('socket' => $socket, 'context' => false, 'header_context' => '', 'body_context' => '', 'method' => 'undefined', 'path' => '', 'version' => '', 'data' => '', 'header' => array(), 'complete' => false, 'createTime' => time(), 'failedCount' => 0,);
     }
 }
